+++ classes/rest/class-stats.php
@@ -10,14 +10,15 @@
  * @package Progress_Planner
  */
 
-namespace Progress_Planner;
+namespace Progress_Planner\Rest;
 
+use Progress_Planner\Base;
 use Progress_Planner\Admin\Widgets\Activity_Scores;
 
 /**
  * Rest_API_Stats class.
  */
-class Rest_API_Stats {
+class Stats {
 	/**
 	 * Constructor.
 	 */
